prepareWitness moved; files renamed
